requirements.txt
# Python dependencies for NewsLens project
# Generated on December 1, 2025
# Python 3.13+ required

# Web Scraping & Data Collection
feedparser==6.0.12
requests==2.31.0
newsapi-python==0.2.7
beautifulsoup4==4.12.2

# NLP & Text Processing
transformers==4.57.3
torch==2.9.1
spacy==3.8.11
nltk==3.8.1
textblob==0.17.1

# Data Processing & Analysis
pandas==2.3.3
numpy==2.3.5

# Database & Storage
sqlalchemy==2.0.44

# Visualization & Dashboard
streamlit==1.51.0
plotly==6.5.0
matplotlib==3.10.7
seaborn==0.13.2
wordcloud==1.9.4

# Configuration & Utilities
pyyaml==6.0.3
python-dotenv==1.2.1

# Logging & Monitoring
loguru==0.7.3

# Testing
pytest==9.0.1
pytest-cov==7.0.0

# Development Tools
black==25.11.0
flake8==7.3.0
